commit work done in branches/tom.brown.code-add_examples
We had 3 scripts that built a feed file, TransitFeed sample in the wiki, table.py and shuttle_from_xmlfeed.py. I added these to a new directory examples/ and created a hack-ish test that runs them. The test also downloads a live copy of the sample in the wiki. The test doesn't not check the output, just that they don't throw any exceptions.