Issue 346:  kmlwriter.py crashes
The current behavior of the kmlwriter is to load the feed, printing any validation messages to the console, and then attempting to run the KML generation step, which often runs into problems if the GTFS doesn't validate properly.  I've updated the code to instead immediately throw an exception if there is a validation error when loading the feed.

Reviewed at http://codereview.appspot.com/6460121/